offcycle: Release circuit provers for v24.2
This is an impromptu PR to release a fork from v17.0.0, adjusted for
v24.2. Whilst this exercise exposes multiple risks, it is done
internally to derisk moving from v24.2 to v25.

This release is unnecessary and is done mostly to carry circuit provers
to mainnet, without being blocked by protocol upgrade.