fix(tee): correct previous fix for race condition in batch locking
Commit a7dc0ed5007f6b2f789f4c61cb3d137843151860 (PR #3342) was supposed
to fix a race condition in batch locking by introducing SQL row-locking,
but it didn't work as expected. Now we are switching back to
coarser-grained table-level locking as [originally suggested][1] by
Harald. The original fix was hard to test unless deployed to `stage` due
to the undeterministic nature of the problem, so we needed to merge it
to the `main` branch to properly test it.

[1]: https://github.com/matter-labs/zksync-era/pull/3342#issuecomment-2514573386