chore: Reduce diff with sync layer stable (zkstack) (#3416)
## What ❔

Introduces functions to the zkstack related to Gateway. These are
included into the libraries, but are not included into the users
interface (CLI itself) and a lot of places contain TODO(EVM-927) to
indicate that those are not usable before the Gateway contracts have
been merged.

Also, as per discussions with @sanekmelnikov it is okay that the first
iteration of gateway scripts (esp since those are not usable by the
partners yet) use println instead of spinners that are more commonly
used in zkstack. For this we have a followup task: EVM-928

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.