fix(api): Propagate fallback errors in traces (#3469)
(hopefully) fixes #3394
See the issue for the context.

Looks like some old transactions do not have error stored for out-of-gas
errors. However, we know that they're failed, and we know that based on
[the error field from the transactions
table](https://github.com/matter-labs/zksync-era/blob/main/core/lib/dal/src/models/storage_transaction.rs#L361).

This PR adds a "fallback" error -- if no other error is detected in the
call trace, we use the error from the sequencer.

Granted, I do not have access to the mainnet DB, so not sure what error
messages are inside, but if I am to guess, it will be more or less
adequate. 😅

⚠️ To make things spicier, I have no way to test this functionality (per
@dutterbutter report in the issue, it does not reproduce with the
current protocol version), so it's kind of "hope it helps" fix. Please
be careful during the review.