refactor: move Cargo.toml from root to core workspace (#3456)
## What ❔

Move `Cargo.toml` from repository root to `core` workspace.

<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

## Why ❔

It is mandatory for proper separation of workspaces and setting up a
`release-please` CI that requires all files of a component to be inside
the component directory.

## CI failures

Currently, the following tests are failing:
* `protobuf_compatibility` - the workflow is using base version and will
be fixed with the next PRs after merge to `main`.

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.