build(consensus): Bump consensus version (#3512)
## What ❔

Bumps version of consensus dependencies to 0.8. New version includes a
needed fix for the consensus algorithm and also removes attesters field
from Genesis spec (as it was already deprecated before).
Tests that use attesters were commented out instead of fixed since most
likely the attester feature will be deprecated when we remove L1 batches
(with ZK OS). But keeping it for now (not sure if we don't want to reuse
the code for something else).