chore: upgrade jsonrpsee and subxt (#3575)
## What ❔

Made sure that we only have one version of jsonrpsee in our dependencies
now and that version is latest (new version is also a bit lighter as
they finally stopped depending on wasm libraries for non-wasm targets).
Had to adapt to breaking changes in `secrecy` as a side effect.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.