fix: block.timestamp is not accurate (#3398)
## What ❔

Related to
https://github.com/zkSync-Community-Hub/zksync-developers/discussions/820

Change the l2 block creation logic to start a new l2 block only when a
transaction is ready to be executed.

## Why ❔

Current logic starts a new l2 block as soon as the previous one is
sealed.

A contract that relies on `block.timestamp` would be able to predict the
time correctly because if the l2 block goes stale (no transaction), then
it will be open indefinitely and the timestamp will not be accurate
anymore

Solution has been tested locally but any feedbacks would be appreciated

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.