feat: add a flag for stage1->stage2 validium migration (#3562)
## What ❔

Introduce an `inclusion_verification_transition_enabled` flag, which is
necessary for enabling inclusion verification in Validiums.
When this flag is used - `da_dispatcher` only populates the older
batches with dummy proofs, and stops polling for inclusion data for new
batches.

The migration would look like this:
- A chain is running with `use_dummy_inclusion_data: true` 
- The following values are updated:
```yaml
da_dispatcher:
  use_dummy_inclusion_data: false
  inclusion_verification_transition_enabled: true
l2:
  da_validator_addr: REAL_VALIDATOR_ADDR
```
- Some period to commit all old batches
- Set `inclusion_verification_transition_enabled: false`

As a side-quest: a check whether dummy inclusion proofs are allowed was
added.

## Why ❔

To ensure that the migration can be performed without downtime.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.