feat(tee): add support for recoverable signatures (#3414)
## What ❔

This PR is part of the effort to implement on-chain TEE proof
verification. Signatures produced by the TEE Prover are now compatible
with the on-chain verifier that uses the `ecrecover` precompile.

## Why ❔

Until now, we've been using _non-recoverable_ signatures in the TEE
prover with a compressed ECDSA public key in each attestation – it was
compressed because there are only 64 bytes available in the report
attestation quote. That worked fine for off-chain proof verification,
but for on-chain verification, it's better to use the Ethereum address
derived from the public key so we can call `ecrecover` in Solidity to
verify the signature.

This PR goes hand in hand with:
- https://github.com/matter-labs/teepot/pull/228
- https://github.com/matter-labs/teepot/pull/240
- https://github.com/matter-labs/teepot/pull/245

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Signed-off-by: Harald Hoyer <harald@matterlabs.dev>
Co-authored-by: Harald Hoyer <harald@matterlabs.dev>