fix(api): Fix pending transactions filter again (#3630)
## What ❔

Changes `received_at` timestamp assignment for transactions, so that
it's reassigned on persistence.

## Why ❔

This makes ordering by `received_at` timestamp causal (transactions with
an earlier timestamp cannot be added to the DB after transactions with a
later timestamp). Causal ordering is currently required for the pending
transactions filter to work correctly.

## Is this a breaking change?

- [ ] Yes
- [x] No

## Operational changes

None.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.