fix(contract-verifier): Correctly process partial verification for EVM contracts (#3688)
I've discovered that partial verification doesn't work correctly for EVM
contracts, because we expect creation bytecode to be equal for compiled
and deployed contracts. This is too strict, because it's OK for them to
be different in metadata hash only.

I didn't add any tests for this (it was somewhat tricky), but checked
manually that the issue was resolved after that.