feat!: Remove old prover stack (#3729)
This PR removes the entire old prover stack. What it means for you:
Firstly, if you're using the deprecated witness_vector_generator &
prover_fri, the binaries are gone. Secondly, you will need to migrate to
circuit_prover (which is faster & comes with extra bells & whistles).

This PR does not attempt to refactor the codebase post removal. It's big
enough as it is, a follow-up cleanup PR will be needed.