fix(data-availability-fetcher): prevent EN database from being populated with unnecessary inclusion data (#3742)
## What ❔

The EN database is populated with the `inclusion_data` received from the
DA layer, not from the main node.
In case the main node has no inclusion data, but the DA layer returns
some - it is being stored in the DB. Later on it is used in
`consistency_checker` which will lead to it failing.

## Why ❔

To prevent inconsistency in `inclusion_data` field that will lead to
`consistency_checker` failing.

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.