fix: make secret optionnal when da client object store is selected (#3715)
## What ❔

Currently if `ObjectStore` is selected as DA client, it requires secret
to be non empty, however there is no secret when using object store.

To bypass this bug, we can add a dummy secret in secrets.yaml otherwise
the server panic such as :

```
da:
  celestia:
    private_key: 0xaa.....

```

This PR solves the issue

## Why ❔

Bug on DA client which can be only fixed with a hack in configuration.
This is not natural and need to be fixed

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes

No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.