fix(contract-verifier): Ignore suppressable errors during compilation (#3747)
Needed for
https://github.com/zkSync-Community-Hub/zksync-developers/discussions/834

For old compiler versions (e.g. zksolc 1.5.4), `suppressedErrors` wasn't
a part of compiler settings, so we cannot easily pass it to the contract
verifier without changing a lot of components. Moreover, it's not a
"real" error but a "heavy warning". If the contract was compiled and
deployed with this warning, we should still allow verifying it.