fix(contract-verifier): handle errors when the contract code is unavailable in Etherscan (#3751)
## What ❔

When a contract is not yet indexed by Etherscan and we try to verify it,
the following error is returned:
`Unable to locate ContractCode at <address>`. This PR adds proper
processing for such errors to give Etherscan time to index the contract.

## Why ❔

Currently, `Unable to locate ContractCode` errors are not properly
processed, and as a result, the verification requests fail.

## Is this a breaking change?
- [ ] Yes
- [X] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [X] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.