feat(consensus): Add consensus protocol versioning (#3720)
## What ❔

It makes some necessary changes to allow supporting multiple consensus
protocol versions. Most of the changes are on the era-consensus side
(which this PR updates to v0.9). On the zksync-era side, I changed the
DB to store versioned block certificates.
This PR also removes the genesis column on the consensus_replica_state
table since that was deprecated a long time ago.
Also updates the Rust toolchain a bit (needed for a stable feature used
in era-consensus). This caused a bunch of lints to need to also be fixed
in this PR (mostly `too_long_first_doc_paragraph`).

Part of BFT-483

## Why ❔

We need to be able to support at least 2 consensus protocol versions
concurrently in order to be able to do consensus protocol upgrades.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes

None.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.