feat: add finality check in DA dispatcher (#3795)
## What ❔

Add a third stage in DA dispatcher. It will ensure the finality of the
blobs that were dispatched in the `dispatch_blob` function. This allows
us to return a softer finality guarantees from the `dispatch_blob`, not
block the execution for a long time, and dispatch more blobs while the
others are being accepted.

This is written with full backward compatibility with EN's `da_fetcher`,
so no changes are needed on the EN side.
Also, the timeout was added to the default Avail client's dispatch
request.

## Why ❔

To make the DA operations more efficient.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.