feat(en): remove dependency on pubdata commitment mode (#3826)
## What ❔

**EN:** 
- Update ConsistencyChecker to not rely on EN config, but rather
determine the pubdata commitment mode based on the batch data
- Add a check into ConsistencyChecker to verify that if Validium mode is
used, `data_availability` must have an entry for this batch
- Update the DA fetcher to handle gaps in `data_availability`, if the
chain was running as a rollup for some time - there will be a range of
batches that don't have the `data_availability` entry

**Core:**
- Update `eth_tx_aggregator` to be able to override the currently
configured `PubdataSendingMode` and `L1BatchCommitmentMode` with the
ones stored in the database.

## Why ❔

To let the node process batches both for Rollup and Validium pubdata
commitment modes to enable Rollup<->Validium transitions.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Co-authored-by: Ivan Schasny <31857042+ischasny@users.noreply.github.com>