refactor(consensus): Deprecate attesters and prepare for validator committee rotation (#3775)
## What ❔

It introduces some code necessary for validator committee rotation. This
doesn't enable that functionality yet as we still need the new consensus
registry contract to be merged (matter-labs/era-contracts#1365) and also
change to matter-labs/era-consensus. A future PR will integrate those
changes and enable it.
It deprecates almost all the code related to the attester role. Since ZK
OS won't have L1 batches, this role won't be needed in the future. Plus,
if we decide to do L1 consensus verification before ZK OS, it's simpler
to just do an operator multisig (the functionality is reduced but it's a
temporary solution anyway).
It also deprecates the genesis column in the consensus_replica_state
Postgres table. Storing genesis has been deprecated in favor of storing
global config a long time ago. #3720 already removed all code related to
it, this completes the deprecation by removing the column.

My next PR will remove the attester method from ZK Stack CLI. I'll do it
in a separate PR since this one is already somewhat large and ZK Stack
CLI changes are best reviewed by someone familiar with it.

Part of PLA-1112 and PLA-1113

## Why ❔

It prepares the repo for validator committee rotation, which is the
ability to change the validator set by interacting with a L2 smart
contract (the consensus registry). This then enables staking, all being
required after it is a contract that handles the staking UI and updates
the validator set on the consensus registry.

## Is this a breaking change?
- [X] Yes
- [ ] No

Technically, yes. Any chain that has a working attester set will stop
attesting batches, but nothing else will break.

## Operational changes
On the consensus config of the main node, the attesters field should be
removed. Not removing it won't break anything, the field will just be
ignored by protobuf.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.