docs: Move interop docs + set create-missing to be false in config (#4020)
## What ❔

<!-- What are the changes this PR brings about? -->
Move interop-related docs to be an independent from bridging section in
docs.
Make `create-missing` be equal to false in mdbook config by default.
This way when the link in any of files is leading to an non-existent
file, mdbook will throw an error instead of creating the file. Earlier
it was turned on by default (at least on my side), but now it's required
to pass it in config. We have to have it to ensure that all links lead
to a valid destinations.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.