feat: private-rpc blacklist of rpc methods (#4048)
Era rpc exposes a lot of methods that require careful handling in the
prividium private rpc. At this stage we are just filtering them out.

## What ❔

Filtering out all methods that might cause to expose more state than
desired in the prividium private rpc.

## Why ❔

In the context of prividium each piece of state requires permisions to
be read. These rpc methods expose state that is not correctly protected,
yet. According to the team discussions we are blacklisting these methods
for now.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.