feat(en): Add revert CLI command for external node (#4053)
## What ❔

Adds a `revert L1_BATCH` command for the EN binary. The command reverts
the node state to the specified batch.

## Why ❔

- This functionality is demanded in practice (e.g., if an EN isn't
updated in time).
- The EN Docker image includes a standalone block_reverter binary, but
using it is inconvenient, given that the way to supply configs to this
binary differs from the EN.

## Is this a breaking change?

- [ ] Yes
- [x] No

## Operational changes

The added command can be used in node maintenance.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.