fix(proof_data_handler): get L1BatchCommitmentMode from database instead of config (#4107)
## What ❔

Get the L1BatchCommitmentMode from the database instead of the config.
This is important during Rollup<->Validium switches, as the config may
not correspond to the commitment mode of the specific batch.

## Why ❔

To eliminate the issues with blobs not being in the witness input for
Rollup batches in case of Rollup->Validium migration or being in the
witness input for Validium batches in case of Validium->Rollup
migration.
When the config change happens during the migration - there might still
be unprocessed batches (or even the unsealed ones) and we need them to
be processed accordingly.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.