fix(da_fetcher): use same logic as in consistency_checker to define a first batch to process (#4322)
## What ❔

Use the logic from a `consistency_checker` to define a first batch to
process. Not only rely on `last_processed_l1_batch`, but also on the
`earliest_l1_batch` and `last_committed_l1_batch`.

## Why ❔

If the EN is synced from a snapshot, but batch 0 doesn't have DA
information in the main node - the EN will never progress because
`da_fetcher` relies on consistency_checker's database table (which is
not updated because there is no DA).

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.