fix(state-keeper): set interop roots on first block in batch (#4324)
## What ❔

We were not properly attaching interop information on the first block in
the batch.
This meant, that if we were sealing only 1 block per batch, interop
roots were never included (this happened during manual tests).

This fix is ignored during protocol upgrades, as this could cause the
batch to fail. The reason for it is that during v29 protocol upgrade,
interop roots cannot be set given that the `L2InteropRootStorage`
contract is not yet deployed.

Additionally, this PR introduces a check to skip importing of interop
roots if settling on the L1, which can cause batch execution to fail as
we do not support L1 interop. This issue was not revealed until after
enabling interop roots on first block in a batch.

---------

Signed-off-by: Danil <deniallugo@gmail.com>
Co-authored-by: zkzoomer <sergio@zkzoomer.io>
Co-authored-by: Danil <deniallugo@gmail.com>