feat: Eth proof manager sender (#4266)
## What ❔

Implementation of EthProofManager component. The component acts as eth
sender/watcher for proofs from proving networks. The component should be
run _alongside_ with proof data handler.
Overall, it works like this:
- Manager picks up the batches with `unpicked` status and tries to send
proof requests to the contract for it. Sending logic is pretty primitive
- we are just sending requests with configurable amount of retries - if
we are not able to send the transactions - we are moving them to prover
cluster.
- Validated proofs are also sent with retries, although, if we weren't
able to send them during this time, we just retry again.
- Component listens for 2 types of events - acknowledgment of the
request/submitted proofs. After receiving a proof, we validate it
locally, for that component needs to know path to verification keys.
- Proof data handler has new config value - proving mode. If set to
PROVER_CLUSTER it works in usual way. If set to PROVING_NETWORK it
operates with a bit different values in the database(this was done
mostly for the sake of backwards compatibility). Also, it runs
additional subcomponent - proof router, which periodically checks the
database and moves batches that(for some reason) are not able to be
proven to prover cluster.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.