fix(upgrades): Read all skipped events  (#4504)
## What ❔

Read upgrades even if we skipped some of them. 

## Why ❔

If we do an upgrade, we have to listen for all events since the latest
upgrade and process them.
E.g. we made an upgrade from v27 to v28.1. And we have to not only check
the upgrade for v28.1 itself, but include v28.0 as well,.

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Signed-off-by: Danil <deniallugo@gmail.com>