feat(eth_sender): Set fusaka upgrade timestamp (#4571)
## What ❔

Use upgrade timestamp as well as block for switching to fusaka blobs

## Why ❔

Unfortunately ETH foundation doesn't provide the exact block for fusaka,
but they have the exact timestamp. So it was decided to use this
timestamp and half an hour before and after this as a safety margin

https://blog.ethereum.org/2025/11/06/fusaka-mainnet-announcement#fusaka-activation

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

Signed-off-by: Danil <deniallugo@gmail.com>